feat(sandbox): shared env vars UI for workspace and pod scopes
